+++ packages/utilities/examples/pagination/discords/AppDiscord.ts
@@ -45,12 +45,12 @@ export abstract class Example {
     }, 25);
 
     const pagination = new Pagination(interaction, embedx, {
-      enableExit: true,
-      onTimeout: (page, message) => {
-        message.delete();
+      ephemeral: true,
+      onTimeout: () => {
+        interaction.deleteReply();
       },
       time: 5 * 1000,
-      type: "SELECT_MENU",
+      type: "BUTTON",
     });
 
     await pagination.send();