+++ packages/utilities/CHANGELOG.md
@@ -19,11 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 > `Security` - in case of vulnerabilities.
 
-## [3.1.0] - 2021-11-08
+## [3.1.1] - 2021-11-08
 
 ### Changed
 
 - renamed `onPaginationTimeout` to `onTimeout`
+
+### Fixed
+
+- If exit mode is enabled for ephermal, throw an error
+- Error when ephermel pagination timed out
 
 ## [3.0.3] - 2021-11-01